# improve nomad ls performance