```
› ls | wc -l
     762
```

before
```
› time nomad ls
nomad ls  8.45s user 6.72s system 99% cpu 15.276 total
```

after
```
› time nomad ls
nomad ls  0.10s user 0.05s system 66% cpu 0.223 total
```